Added: Teleport to Start menu option, monsters always drop egg game option, monster shift appearance option, and exp multiplier option.
Monster locations are now cached on connection to avoid an async bug that was causing some monster groups to not load their AI correctly. I still occasionally get a NullReferenceException when scouting a location, but it doesn't seem to break anything. Continuing to monitor to see if it does eventually break something.